### 2024-01-13 #6228 @editorialbot
[PRE REVIEW]: Fusilli: A Python package for multimodal data fusion
**Submitting author:** @florencejt (<a href="http://orcid.org/0000-0001-7803-5682">Florence J Townend</a>)
**Repository:** https://github.com/florencejt/fusilli
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.2
