### 2024-01-15 #6228 @florencejt
Hi @arfon, thank you for letting me know!
I have one question about the submission/review process: is it permitted to release new tagged versions of the software during the JOSS process or is it best to wait until the whole process is completed before tinkering?
