### 2024-01-31 af26d5f
small bug fix to data loading to make it able to only have one tabular data source specified
