### 2024-01-31 7a2cffd
Added input arguments to traintestdatamodule and traintestgraphdatamodule to specify own test indices
