### 2024-01-31 4a9948a
added test for traintestdatamodule using specified test indices
