### 2024-01-31 a225077
added test for the traintestgraphdatamodule
