### 2024-01-31 f789c83
added setting own kfold indices to kfolddatamodule and wrote test
