### 2024-01-31 022569c
separated out kfold and train test data module function call in prepare_fusion_data so we dont need to specify test size for kfold and num folds for train test redundantly
