### 2024-01-31 0f5fdb8
added train test customisability to documentation on customising training
