### 2024-02-26 #6414 @editorialbot
[REVIEW]: Fusilli: A Python package for multimodal data fusion
**Submitting author:** @florencejt (<a href="http://orcid.org/0000-0001-7803-5682">Florence J Townend</a>)
**Repository:** https://github.com/florencejt/fusilli
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.2
