### 2024-03-25 #6414 @atrisovic
Hi @aaronhan223, @felixkrones, hope you're having a good day! Just a friendly reminder about the review for Fusilli: A Python package for multimodal data fusion. Your expertise is invaluable to this process, and we're looking forward to your next steps. Please update us on your progress when you can. Thanks so much!
