### 2024-04-15 7462dbe
changing model outputs for concat data, mcvae, and activation to be single numbers for experimenting with shap. Also experimented with a list input for ConcatData
