+++ fusilli/fusionmodels/tabularfusion/activation.py
@@ -11,9 +11,9 @@ from fusilli.utils import check_model_validity
 
 class ActivationFusion(ParentFusionModel, nn.Module):
     """
-    Performs an element wise product of the feature maps of the two tabular modalities, 
+    Performs an element wise product of the feature maps of the two tabular modalities,
     tanh activation function and sigmoid activation function. Afterwards the the first tabular modality feature
-    map is concatenated with the fused feature map. 
+    map is concatenated with the fused feature map.
 
     Attributes
     ----------
@@ -55,7 +55,9 @@ class ActivationFusion(ParentFusionModel, nn.Module):
         multiclass_dimensions : int
             Number of classes in the multiclass classification problem.
         """
-        ParentFusionModel.__init__(self, prediction_task, data_dims, multiclass_dimensions)
+        ParentFusionModel.__init__(
+            self, prediction_task, data_dims, multiclass_dimensions
+        )
 
         self.prediction_task = prediction_task
 
@@ -94,7 +96,8 @@ class ActivationFusion(ParentFusionModel, nn.Module):
         mod2_output_dim = list(self.mod2_layers.values())[-1][0].out_features
         if mod1_output_dim != mod2_output_dim:
             raise UserWarning(
-                "The number of output features of mod1_layers and mod2_layers must be the same for Activation fusion. Please change the final layers in the modality layers to have the same number of output features as each other.")
+                "The number of output features of mod1_layers and mod2_layers must be the same for Activation fusion. Please change the final layers in the modality layers to have the same number of output features as each other."
+            )
 
         self.get_fused_dim()
         self.fused_layers, out_dim = check_model_validity.check_fused_layers(
@@ -145,9 +148,7 @@ class ActivationFusion(ParentFusionModel, nn.Module):
 
         out = self.final_prediction(out_fuse)
 
-        return [
-            out,
-        ]
+        return out
 
 
 """
