+++ fusilli/fusionmodels/tabularfusion/mcvae_model.py
@@ -136,7 +136,9 @@ class MCVAESubspaceMethod:
         init_dict = {
             "n_channels": 2,
             "lat_dim": self.num_latent_dims,
-            "n_feats": tuple([self.datamodule.data_dims[0], self.datamodule.data_dims[1]]),
+            "n_feats": tuple(
+                [self.datamodule.data_dims[0], self.datamodule.data_dims[1]]
+            ),
         }
 
         self.fit_model = Mcvae(**init_dict, sparse=True)
@@ -261,7 +263,9 @@ class MCVAESubspaceMethod:
         with contextlib.redirect_stdout(None):
             mcvae_fit.optimize(epochs=self.max_epochs, data=mcvae_training_data)
             ideal_epoch = mcvae_early_stopping_tol(
-                tolerance=mcvae_tolerance, patience=mcvae_patience, loss_logs=mcvae_fit.loss["total"]
+                tolerance=mcvae_tolerance,
+                patience=mcvae_patience,
+                loss_logs=mcvae_fit.loss["total"],
             )
 
         mcvae_esfit = Mcvae(**init_dict, sparse=True)
@@ -284,7 +288,9 @@ class MCVAESubspaceMethod:
         # getting mean latent space
         mean_latents = self.get_latents(mcvae_training_data)
 
-        return torch.Tensor(mean_latents), pd.DataFrame(labels, columns=["prediction_label"])
+        return torch.Tensor(mean_latents), pd.DataFrame(
+            labels, columns=["prediction_label"]
+        )
 
     def convert_to_latent(self, test_dataset):
         """
@@ -373,7 +379,9 @@ class MCVAE_tab(ParentFusionModel, nn.Module):
         multiclass_dimensions : int
             Number of classes in the multiclass classification task.
         """
-        ParentFusionModel.__init__(self, prediction_task, data_dims, multiclass_dimensions)
+        ParentFusionModel.__init__(
+            self, prediction_task, data_dims, multiclass_dimensions
+        )
 
         self.prediction_task = prediction_task
 
@@ -452,6 +460,4 @@ class MCVAE_tab(ParentFusionModel, nn.Module):
 
         out_pred = self.final_prediction(out_fuse)
 
-        return [
-            out_pred,
-        ]
+        return out_pred
