### 2024-04-15 89e205d
changed all models to have x1 and x2 inputs instead of tensors, and only one output instead of lists so that shap values can be used
