+++ fusilli/fusionmodels/tabularfusion/concat_data.py
@@ -86,7 +86,7 @@ class ConcatTabularData(ParentFusionModel, nn.Module):
 
         self.set_final_pred_layers(out_dim)
 
-    def forward(self, x):
+    def forward(self, x1, x2):
         """
         Forward pass of the model.
 
@@ -97,11 +97,11 @@ class ConcatTabularData(ParentFusionModel, nn.Module):
 
         Returns
         -------
-        list
-            List containing the output of the model.
+        torch.Tensor
+            Prediction of the model.
         """
 
-        print("List input", isinstance(x, list))
+        # print("List input", isinstance(x, list))
         # print("x data type", type(x))
         # print("x length", len(x))
 
@@ -110,7 +110,7 @@ class ConcatTabularData(ParentFusionModel, nn.Module):
         # ~~ Checks ~~
         # check_model_validity.check_model_input(x)
 
-        x_fuse = torch.cat(x, -1)
+        x_fuse = torch.cat((x1, x2), -1)
 
         out_fuse = self.fused_layers(x_fuse)
 
