+++ tests/test_models/test_BaseModel.py
@@ -17,7 +17,7 @@ class SampleGraphFusionModel(nn.Module):
         self.graph_maker = Mock()
 
     def forward(self, x):
-        return [torch.rand((x[0].shape[0], 1)), ]
+        return torch.rand((x[0].shape[0], 1))
 
 
 class SampleFusionModel(nn.Module):
@@ -28,11 +28,8 @@ class SampleFusionModel(nn.Module):
         self.fusion_type = "attention"  # Sample fusion_type
         self.subspace_method = None
 
-    def forward(self, x):
-        if isinstance(x, tuple):
-            return [torch.rand((x[0].shape[0], 1)), ]
-        else:
-            return [torch.rand((x.shape[0], 1)), ]
+    def forward(self, x1, x2):
+        return torch.rand((x1.shape[0], 1))
 
 
 class SampleFusionModelReconstructions(nn.Module):
@@ -44,11 +41,8 @@ class SampleFusionModelReconstructions(nn.Module):
         self.subspace_method = None
         self.custom_loss = Mock(return_value=0.5)
 
-    def forward(self, x):
-        if isinstance(x, tuple):
-            return [torch.rand((x[0].shape[0], 1)), torch.rand(x[0].shape)]
-        else:
-            return [torch.rand((x.shape[0], 1)), torch.rand(x.shape)]
+    def forward(self, x1, x2):
+        return [torch.rand((x1.shape[0], 1)), torch.rand(x2.shape)]
 
 
 @pytest.fixture
@@ -111,23 +105,36 @@ def test_get_data_from_batch(sample_model):
     assert y.shape == torch.Size([10])
 
     # more than 2 modalities
-    batch = (torch.rand((10, 12)), torch.rand((10, 15)), torch.rand((10, 18)), torch.randint(1, (10,)))
-    with pytest.raises(ValueError, match=re.escape("Batch size is not 2 (preds and labels) or 3 (2 pred data types "
-                                                   "and "
-                                                   "labels) modalities long")):
+    batch = (
+        torch.rand((10, 12)),
+        torch.rand((10, 15)),
+        torch.rand((10, 18)),
+        torch.randint(1, (10,)),
+    )
+    with pytest.raises(
+        ValueError,
+        match=re.escape(
+            "Batch size is not 2 (preds and labels) or 3 (2 pred data types "
+            "and "
+            "labels) modalities long"
+        ),
+    ):
         x, y = model.get_data_from_batch(batch)
 
 
 def test_get_data_from_batch_graph(sample_graph_model):
     model = sample_graph_model
 
-    class GraphDataClass():
+    class GraphDataClass:
         def __init__(self, x, y):
             self.x = x
             self.y = y
-            self.edge_index = torch.tensor([[0, 1, 1, 2],
-                                            [1, 0, 2, 1]], dtype=torch.long)
-            self.edge_attr = torch.tensor([[1.0], [2.0], [3.0], [4.0]], dtype=torch.float)
+            self.edge_index = torch.tensor(
+                [[0, 1, 1, 2], [1, 0, 2, 1]], dtype=torch.long
+            )
+            self.edge_attr = torch.tensor(
+                [[1.0], [2.0], [3.0], [4.0]], dtype=torch.float
+            )
 
     batch = GraphDataClass(torch.rand((12, 10)), torch.randint(1, (10,)))
 
@@ -160,20 +167,19 @@ def test_get_model_outputs(sample_model, sample_model_recon):
 
     # with reconstructions
     model2 = sample_model_recon
-    preds, reconstructions = model2.get_model_outputs(batch)
[…]
