+++ fusilli/fusionmodels/tabularfusion/activation.py
@@ -107,42 +107,42 @@ class ActivationFusion(ParentFusionModel, nn.Module):
         # setting final prediction layers with final out features of fused layers
         self.set_final_pred_layers(out_dim)
 
-    def forward(self, x):
+    def forward(self, x1, x2):
         """
         Forward pass of the model.
 
         Parameters
         ----------
-        x : tuple
-            Tuple containing the input data.
+        x1 : torch.Tensor
+            Input tensor of the first modality.
+        x2 : torch.Tensor
+            Input tensor of the second modality.
 
         Returns
         -------
-        list
-            List containing the output of the model.
+        out : torch.Tensor
+            Fused prediction.
         """
 
         # ~~ Checks ~~
-        check_model_validity.check_model_input(x)
-
-        x_tab1 = x[0]
-        x_tab2 = x[1]
+        check_model_validity.check_model_input(x1)
+        check_model_validity.check_model_input(x2)
 
         for layer in self.mod1_layers.values():
-            x_tab1 = layer(x_tab1)
+            x1 = layer(x1)
 
         for layer in self.mod2_layers.values():
-            x_tab2 = layer(x_tab2)
+            x2 = layer(x2)
 
-        x_tab1 = torch.squeeze(x_tab1, 1)
-        x_tab2 = torch.squeeze(x_tab2, 1)
+        x1 = torch.squeeze(x1, 1)
+        x2 = torch.squeeze(x2, 1)
 
-        out_fuse = torch.mul(x_tab1, x_tab2)
+        out_fuse = torch.mul(x1, x2)
 
         out_fuse = torch.tanh(out_fuse)
         out_fuse = torch.sigmoid(out_fuse)
 
-        out_fuse = torch.cat((out_fuse, x_tab1), dim=1)
+        out_fuse = torch.cat((out_fuse, x1), dim=1)
 
         out_fuse = self.fused_layers(out_fuse)
 
