+++ fusilli/fusionmodels/tabularfusion/mcvae_model.py
@@ -445,11 +445,11 @@ class MCVAE_tab(ParentFusionModel, nn.Module):
 
         Returns
         -------
-        out_pred : list
-            List containing the predictions.
+        torch.Tensor
+            Tensor containing the predicted values.
         """
         # ~~ Checks ~~
-        check_model_validity.check_model_input(x, True)
+        check_model_validity.check_model_input(x)
 
         x_latent = x
 
