### 2024-04-18 8d58e9f
making branch for adding more modalities to the methods
