+++ docs/examples/customising_behaviour/README.rst
@@ -1,4 +1,4 @@
-.. _advanced-examples:
+.. _customising-examples:
 
 Customising Fusilli
 =======================================================
