### 2024-04-19 ac6d9a2
changed data loading to account for 3 tabular modalities and turned data dims into a dictionary instead of a list
