+++ fusilli/data.py
@@ -99,6 +99,10 @@ class CustomDataset(Dataset):
         Tensor of predictive features for modality 1.
     dataset2 : tensor
         Tensor of predictive features for modality 2.
+    three_modalities : bool
+        Flag for three modalities. True if three modalities, False if two modalities.
+    dataset3 : tensor
+        Tensor of predictive features for modality 3 if three modalities.
     dataset : tensor
         Tensor of predictive features for uni-modal data.
     labels : tensor
@@ -128,8 +132,16 @@ class CustomDataset(Dataset):
         # only 2 modalities are supported currently
         if isinstance(pred_features, list):
             self.multimodal_flag = True
+
             self.dataset1 = pred_features[0].float()
             self.dataset2 = pred_features[1].float()
+
+            # if there is a third tabular modality
+            if len(pred_features) > 2:
+                self.three_modalities = True
+                self.dataset3 = pred_features[2].float()
+            else:
+                self.three_modalities = False
 
         # if pred_features is a tensor: it's unimodal data
         elif isinstance(pred_features, torch.Tensor):
@@ -173,7 +185,15 @@ class CustomDataset(Dataset):
             Index of the item to return.
         """
         if self.multimodal_flag:
-            return self.dataset1[idx], self.dataset2[idx], self.labels[idx]
+            if self.three_modalities:
+                return (
+                    self.dataset1[idx],
+                    self.dataset2[idx],
+                    self.dataset3[idx],
+                    self.labels[idx],
+                )
+            else:
+                return self.dataset1[idx], self.dataset2[idx], self.labels[idx]
         else:
             return self.dataset[idx], self.labels[idx]
 
@@ -200,9 +220,8 @@ class LoadDatasets:
         """
         Parameters
         ----------
-        sources : list
-            List of source csv files.
-            [tabular1_source, tabular2_source, img_source]
+        sources : dict
+            Dictionary of source csv files with keys the possible keys "tabular1", "tabular2", "tabular3", and "image".
         img_downsample_dims : tuple
             Size to downsample the images to (height, width, depth) or (height, width) for 2D
             images.
@@ -217,26 +236,49 @@ class LoadDatasets:
             "ID".
 
         """
-        self.tabular1_source, self.tabular2_source, self.img_source = sources
+
+        # get the source files into attributes, returns empty string if not present
+        self.tabular1_source = sources.get("tabular1", "")
+        self.tabular2_source = sources.get("tabular2", "")
+        self.tabular3_source = sources.get("tabular3", "")
+        self.img_source = sources.get("image", "")
+
+        # set the image downsample size
         self.image_downsample_size = (
             img_downsample_dims  # can choose own image size here
         )
 
-        # read in the csv files and raise errors if they don't have the right columns
-        # or if the index column is not named "ID"
-        tab1_df = pd.read_csv(self.tabular1_source)
-        if "ID" not in tab1_df.columns:
-            raise ValueError("The CSV must have an index column named 'ID'.")
-        if "prediction_label" not in tab1_df.columns:
-            raise ValueError("The CSV must have a label column named 'prediction_label'.")
+        def check_csv_columns(source_path):
+            """
+            Checks the columns of the CSV files.
 
[…]

[…]
