### 2024-04-19 3102a0a
TabularDecision working with 3 tabular modalities
