+++ fusilli/fusionmodels/tabularfusion/channelwise_att.py
@@ -2,6 +2,11 @@
 Channel-wise multiplication fusion model for tabular data.
 """
 
+# TODO make 3-tabular data work
+
+# TODO add argument to say which tabular modality is the main one
+
+
 import torch.nn as nn
 from fusilli.fusionmodels.base_model import ParentFusionModel
 from fusilli.utils import check_model_validity
