+++ fusilli/fusionmodels/tabularfusion/mcvae_model.py
@@ -3,6 +3,8 @@ This module implements the MCVAE (multi-channel variational autoencoder) model f
 two types of tabular data.
 """
 
+# TODO make 3-tabular data work
+
 import torch.nn as nn
 from fusilli.fusionmodels.base_model import ParentFusionModel
 import torch
