### 2024-04-19 2305272
altered concat tabular to work with three modalities
