### 2024-04-19 ff586ca
made concat_feature_maps tabular work with three tabular modalities
