+++ fusilli/fusionmodels/tabularimagefusion/concat_img_maps_tabular_maps.py
@@ -50,8 +50,8 @@ class ConcatImageMapsTabularMaps(ParentFusionModel, nn.Module):
         ----------
         prediction_task : str
             Type of prediction to be performed.
-        data_dims : list
-            List containing the dimensions of the data.
+        data_dims : dict
+            Dictionary of data dimensions with keys "mod1_dim", "mod2_dim", "mod3_dim", and "img_dim".
         multiclass_dimensions : int
             Number of classes in the multiclass classification task.
         """
