### 2024-04-22 8e8df72
changed tabular channelwise attention to work with three modalities and also made it possible to use the >main modality< so to speak
