+++ fusilli/fusionmodels/tabularfusion/mcvae_model.py
@@ -3,16 +3,20 @@ This module implements the MCVAE (multi-channel variational autoencoder) model f
 two types of tabular data.
 """
 
-# TODO make 3-tabular data work
-
 import torch.nn as nn
 from fusilli.fusionmodels.base_model import ParentFusionModel
 import torch
-from fusilli.utils.mcvae.src.mcvae.models import Mcvae
+
+# from fusilli.utils.mcvae.src.mcvae.models import Mcvae
+
+# FOR LOCAL INSTALL - CHANGE WHEN PUSHING TO RELEASE
+from mcvae.models import Mcvae
+
 import contextlib
 import pandas as pd
 import numpy as np
 from fusilli.utils.training_utils import get_checkpoint_filenames_for_subspace_models
+import sys
 
 from fusilli.utils import check_model_validity
 
@@ -135,15 +139,28 @@ class MCVAESubspaceMethod:
         new_checkpoint_path = checkpoint_path[0][: -len(".ckpt")]
         checkpoint = torch.load(new_checkpoint_path)
 
-        init_dict = {
-            "n_channels": 2,
-            "lat_dim": self.num_latent_dims,
-            "n_feats": tuple(
+        if self.datamodule.data_dims["mod3_dim"] is not None:
+            n_channels = 3
+            n_features = tuple(
+                [
+                    self.datamodule.data_dims["mod1_dim"],
+                    self.datamodule.data_dims["mod2_dim"],
+                    self.datamodule.data_dims["mod3_dim"],
+                ]
+            )
+        else:
+            n_channels = 2
+            n_features = tuple(
                 [
                     self.datamodule.data_dims["mod1_dim"],
                     self.datamodule.data_dims["mod2_dim"],
                 ]
-            ),
+            )
+
+        init_dict = {
+            "n_channels": n_channels,
+            "lat_dim": self.num_latent_dims,
+            "n_feats": n_features,
         }
 
         self.fit_model = Mcvae(**init_dict, sparse=True)
@@ -176,31 +193,44 @@ class MCVAESubspaceMethod:
         Parameters
         ----------
         dataset : list
-            List containing the two types of tabular data.
+            List containing the types of tabular data.
 
         Returns
         -------
         mean_latents : np.array
             Array containing the mean latents of the dataset.
         """
-        # getting mean latent space
+
+        # how many channels are in this method (2 or 3, equal to number of tabular modalities)
+        num_channels = len(dataset)
 
         q = self.fit_model.encode(dataset)
 
         latent_vars_ch0 = q[0].loc.detach().cpu()
         latent_vars_ch1 = q[1].loc.detach().cpu()
+        if num_channels == 3:
+            latent_vars_ch2 = q[2].loc.detach().cpu()
         latents = []
 
         n_dims = latent_vars_ch0.shape[1]
 
         for i in range(n_dims):
-            latent_temp = np.vstack([latent_vars_ch0[:, i], latent_vars_ch1[:, i]])
+            if num_channels == 3:
+                latent_temp = np.vstack(
+                    [
+                        latent_vars_ch0[:, i],
+                        latent_vars_ch1[:, i],
+                        latent_vars_ch2[:, i],
+                    ]
+                )
+            else:
+                latent_temp = np.vstack([latent_vars_ch0[:, i], latent_vars_ch1[:, i]])
             latents.append(np.mean(latent_temp, axis=0))
 
         indices = [i for i in range(self.num_latent_dims)]
         latents = [latents[i] for i in indices]
 
-        mean_latents = np.vstack([latents]).transpose()  # 43 people
[…]
