+++ fusilli/fusionmodels/tabularfusion/channelwise_att.py
@@ -2,11 +2,6 @@
 Channel-wise multiplication fusion model for tabular data.
 """
 
-# TODO make 3-tabular data work
-
-# TODO add argument to say which tabular modality is the main one
-
-
 import torch.nn as nn
 from fusilli.fusionmodels.base_model import ParentFusionModel
 from fusilli.utils import check_model_validity
@@ -48,14 +43,14 @@ class TabularChannelWiseMultiAttention(ParentFusionModel, nn.Module):
       Dictionary containing the layers of the 1st type of tabular data.
     mod2_layers : nn.ModuleDict
       Dictionary containing the layers of the 2nd type of tabular data.
-    match_dim_layers : nn.ModuleDict
-      Module dictionary containing the linear layers to make the dimensions of the two types of
-      tabular data the same. This is done to ensure that the channel-wise multiplication can be
-      performed. This doesn't change the mod1_layers or mod2_layers, it just makes the outputs
-      multipliable.
+    mod3_layers : nn.ModuleDict (optional)
+        Dictionary containing the layers of the 3rd type of tabular data. If 3 tabular data are not provided, this is not used.
+    match_dim_layers: dict
+        Dictionary containing module dictionaries of linear layers to make the dimensions of the types of tabular data the same at each layer.
+        This is done to ensure that the channel-wise multiplication can be performed and does not alter the original layers.
     fused_dim : int
       Number of features of the fused layers. This is the output size of the
-      2nd type of tabular data's layers.
+      final main tabular data's layers.
     fused_layers : nn.Sequential
       Sequential layer containing the fused layers.
     final_prediction : nn.Sequential
@@ -69,6 +64,8 @@ class TabularChannelWiseMultiAttention(ParentFusionModel, nn.Module):
     modality_type = "tabular_tabular"
     #: str: Type of fusion.
     fusion_type = "attention"
+    #: str: Available for three tabular modalities.
+    three_modalities = True
 
     def __init__(self, prediction_task, data_dims, multiclass_dimensions):
         """
@@ -79,7 +76,7 @@ class TabularChannelWiseMultiAttention(ParentFusionModel, nn.Module):
         data_dims : list
           List containing the dimensions of the data.
         multiclass_dimensions : int
-            Number of classes in the dataset.
+                Number of classes in the dataset.
         """
         ParentFusionModel.__init__(
             self, prediction_task, data_dims, multiclass_dimensions
@@ -89,6 +86,11 @@ class TabularChannelWiseMultiAttention(ParentFusionModel, nn.Module):
 
         self.set_mod1_layers()
         self.set_mod2_layers()
+        if self.data_dims["mod3_dim"] is not None:
+            self.set_mod3_layers()
+
+        # Which modality is the main one?
+        self.main_modality = 1  # 1, 2, or 3 depending on which modality is the main one
 
         self.get_fused_dim()
         self.set_fused_layers(self.fused_dim)
@@ -103,11 +105,21 @@ class TabularChannelWiseMultiAttention(ParentFusionModel, nn.Module):
         -------
         None.
         """
-        self.fused_dim = list(self.mod2_layers.values())[-1][0].out_features
+
+        if self.main_modality == 1:
+            self.fused_dim = list(self.mod1_layers.values())[-1][0].out_features
+        elif self.main_modality == 2:
+            self.fused_dim = list(self.mod2_layers.values())[-1][0].out_features
+        elif self.main_modality == 3:
+            self.fused_dim = list(self.mod3_layers.values())[-1][0].out_features
+        else:
+            raise ValueError("main_modality must be 1, 2, or 3")
+
+        # self.fused_dim = list(self.mod2_layers.values())[-1][0].out_features
 
     def calc_fused_layers(self):
         """
-        Calculates the fusion layers.
[…]
