### 2024-04-22 acd5762
added tests for three modality functionality in test_tabularfusion and picked up on some data_dims lists I didnt notice before
