+++ tests/test_models/test_tabularfusion.py
@@ -30,9 +30,13 @@ def test_ConcatTabularFeatureMaps():
     assert test_model.modality_type == "tabular_tabular"
     assert hasattr(test_model, "fusion_type")
     assert test_model.fusion_type == "operation"
+    assert hasattr(test_model, "three_modalities")
+    assert test_model.three_modalities == True
 
     test_model = test_model(
-        prediction_task="binary", data_dims=[10, 14, None], multiclass_dimensions=None
+        prediction_task="binary",
+        data_dims={"mod1_dim": 10, "mod2_dim": 14, "mod3_dim": None, "img_dim": None},
+        multiclass_dimensions=None,
     )
 
     # initialising
@@ -60,6 +64,42 @@ def test_ConcatTabularFeatureMaps():
     assert isinstance(test_output, torch.Tensor)
     assert test_output.shape == torch.Size([8, 1])
 
+    # Three tabular modalities
+    test_model = fusion_model_dict["ConcatTabularFeatureMaps"]
+    test_model = test_model(
+        prediction_task="binary",
+        data_dims={"mod1_dim": 10, "mod2_dim": 14, "mod3_dim": 20, "img_dim": None},
+        multiclass_dimensions=None,
+    )
+
+    # initialising
+    assert isinstance(test_model, nn.Module)
+    assert isinstance(test_model, ParentFusionModel)
+    assert hasattr(test_model, "prediction_task")
+    assert test_model.prediction_task == "binary"
+    assert hasattr(test_model, "mod1_layers")
+    assert test_model.mod1_layers["layer 1"][0].in_features == 10
+    assert hasattr(test_model, "mod2_layers")
+    assert test_model.mod2_layers["layer 1"][0].in_features == 14
+    assert hasattr(test_model, "mod3_layers")
+    assert test_model.mod3_layers["layer 1"][0].in_features == 20
+    assert hasattr(test_model, "fused_dim")
+    assert (
+        test_model.fused_dim
+        == test_model.mod1_layers["layer 5"][0].out_features
+        + test_model.mod2_layers["layer 5"][0].out_features
+        + test_model.mod3_layers["layer 5"][0].out_features
+    )
+    assert hasattr(test_model, "fused_layers")
+    assert hasattr(test_model, "final_prediction")
+    assert hasattr(test_model, "forward")
+
+    # forward pass
+    test_input = (torch.randn(8, 10), torch.randn(8, 14), torch.randn(8, 20))
+    test_output = test_model.forward(test_input[0], test_input[1], test_input[2])
+    assert isinstance(test_output, torch.Tensor)
+    assert test_output.shape == torch.Size([8, 1])
+
 
 # fusilli.fusionmodels.tabularfusion.concat_data.ConcatTabularData
 def test_ConcatTabularData():
@@ -72,9 +112,13 @@ def test_ConcatTabularData():
     assert test_model.modality_type == "tabular_tabular"
     assert hasattr(test_model, "fusion_type")
     assert test_model.fusion_type == "operation"
+    assert hasattr(test_model, "three_modalities")
+    assert test_model.three_modalities == True
 
     test_model = test_model(
-        prediction_task="binary", data_dims=[10, 14, None], multiclass_dimensions=None
+        prediction_task="binary",
+        data_dims={"mod1_dim": 10, "mod2_dim": 14, "mod3_dim": None, "img_dim": None},
+        multiclass_dimensions=None,
     )
 
     # initialising
@@ -97,11 +141,41 @@ def test_ConcatTabularData():
     assert isinstance(test_output, torch.Tensor)
     assert test_output.shape == torch.Size([8, 1])
 
+    # Three tabular modalities
[…]
