### 2024-04-22 928e8c5
Made crossmodal attention possible with three tabular modalities. Also altered the method not concatenate the unimodal feature maps at the end because that's not in the original paper
