+++ fusilli/fusionmodels/tabularfusion/crossmodal_att.py
@@ -64,6 +64,8 @@ class TabularCrossmodalMultiheadAttention(ParentFusionModel, nn.Module):
     modality_type = "tabular_tabular"
     #: str: Type of fusion.
     fusion_type = "attention"
+    #: str: Available for three tabular modalities.
+    three_modalities = True
 
     def __init__(self, prediction_task, data_dims, multiclass_dimensions):
         """
@@ -85,6 +87,9 @@ class TabularCrossmodalMultiheadAttention(ParentFusionModel, nn.Module):
 
         self.set_mod1_layers()
         self.set_mod2_layers()
+        if self.data_dims["mod3_dim"] is not None:
+            self.set_mod3_layers()
+
         self.calc_fused_layers()
 
         self.relu = nn.ReLU()
@@ -105,32 +110,58 @@ class TabularCrossmodalMultiheadAttention(ParentFusionModel, nn.Module):
             If dtype of the layers is not nn.ModuleDict.
 
         """
-        # if mod1 and mod2 have a different number of layers, return error
         check_model_validity.check_dtype(self.mod1_layers, nn.ModuleDict, "mod1_layers")
         check_model_validity.check_dtype(self.mod2_layers, nn.ModuleDict, "mod2_layers")
 
-        if len(self.mod1_layers) != len(self.mod2_layers):
-            raise ValueError(
-                "The number of layers in the two modalities must be the same."
+        # if we've got a 3rd tabular modality
+        if self.data_dims["mod3_dim"] is not None:
+
+            # check the layers are valid (if they've been changed)
+            check_model_validity.check_dtype(
+                self.mod3_layers, nn.ModuleDict, "mod3_layers"
             )
 
+            # check that the number of layers in each modality is the same
+            if (
+                len(self.mod1_layers) != len(self.mod2_layers)
+                or len(self.mod1_layers) != len(self.mod3_layers)
+                or len(self.mod2_layers) != len(self.mod3_layers)
+            ):
+                raise ValueError(
+                    "The number of layers in the three modalities must be the same."
+                )
+        else:
+            # check that the number of layers in each modality is the same if we don't have a 3rd modality
+            if len(self.mod1_layers) != len(self.mod2_layers):
+                raise ValueError(
+                    "The number of layers in the two modalities must be the same."
+                )
+
+        # Output size of the first tabular modality
         self.fused_dim = list(self.mod1_layers.values())[-1][0].out_features
 
-        # self.set_fused_layers(self.fused_dim)
+        mod1_output_size = list(self.mod1_layers.values())[-1][0].out_features
+        mod2_output_size = list(self.mod2_layers.values())[-1][0].out_features
 
-        self.tab1_to_embed_dim = nn.Linear(self.fused_dim, self.attention_embed_dim)
-        self.tab2_to_embed_dim = nn.Linear(
-            list(self.mod2_layers.values())[-1][0].out_features,
-            self.attention_embed_dim,
-        )
+        self.tab1_to_embed_dim = nn.Linear(mod1_output_size, self.attention_embed_dim)
+        self.tab2_to_embed_dim = nn.Linear(mod2_output_size, self.attention_embed_dim)
+
+        if self.data_dims["mod3_dim"] is not None:
+            mod3_output_size = list(self.mod3_layers.values())[-1][0].out_features
+            self.tab3_to_embed_dim = nn.Linear(
+                mod3_output_size, self.attention_embed_dim
+            )
 
         self.attention = nn.MultiheadAttention(
[…]
