### 2024-04-22 9a48ffa
altered activation fusion to work with 3 modalities and able to choose which modality is the >main< modality
