+++ fusilli/fusionmodels/tabularfusion/activation.py
@@ -24,12 +24,18 @@ class ActivationFusion(ParentFusionModel, nn.Module):
     ----------
     prediction_task : str
         Type of prediction to be performed.
+    main_modality : int
+        Which modality is the main one? 1, 2, or 3 depending on which modality is the main one.
+        The main modality has its final feature map concatenated with the fused feature map.
+        Default is 1.
     mod1_layers : nn.ModuleDict
         Dictionary containing the layers of the first modality. Calculated in the
         :meth:`~ParentFusionModel.set_mod1_layers` method.
     mod2_layers : nn.ModuleDict
         Dictionary containing the layers of the second modality. Calculated in the
         :meth:`~ParentFusionModel.set_mod2_layers` method.
+    mod3_layers : nn.ModuleDict (optional)
+        Dictionary containing the layers of the 3rd type of tabular data. If 3 tabular data are not provided, this is not used.
     fused_dim : int
         Number of features of the fused layers. In this method, it's the size of the tabular 1
         layers output plus the size of the tabular 2 layers output.
@@ -48,6 +54,8 @@ class ActivationFusion(ParentFusionModel, nn.Module):
     modality_type = "tabular_tabular"
     #: str: Type of fusion.
     fusion_type = "operation"
+    #: str: Available for three tabular modalities.
+    three_modalities = True
 
     def __init__(self, prediction_task, data_dims, multiclass_dimensions):
         """
@@ -69,6 +77,12 @@ class ActivationFusion(ParentFusionModel, nn.Module):
         self.set_mod1_layers()
         self.set_mod2_layers()
 
+        if self.data_dims["mod3_dim"] is not None:
+            self.set_mod3_layers()
+
+        # Which modality is the main one?
+        self.main_modality = 1  # 1, 2, or 3 depending on which modality is the main one
+
         self.get_fused_dim()
         self.set_fused_layers(self.fused_dim)
 
@@ -81,7 +95,11 @@ class ActivationFusion(ParentFusionModel, nn.Module):
         """
         mod1_output_dim = list(self.mod1_layers.values())[-1][0].out_features
         mod2_output_dim = list(self.mod2_layers.values())[-1][0].out_features
-        # New fused dimension is the sum of mod1 and mod2 output dimensions
+        if self.data_dims["mod3_dim"] is not None:
+            mod3_output_dim = list(self.mod3_layers.values())[-1][0].out_features
+
+        # New fused dimension is the sum of the output feature map dimensions of one of the modalities (doesn't matter which one because they are the same)
+        # And adding the feature map dimension of the "main" modality, again, doesn't matter which one because they are the same
         self.fused_dim = mod1_output_dim + mod2_output_dim
 
     def calc_fused_layers(self):
@@ -101,8 +119,19 @@ class ActivationFusion(ParentFusionModel, nn.Module):
         mod2_output_dim = list(self.mod2_layers.values())[-1][0].out_features
         if mod1_output_dim != mod2_output_dim:
             raise UserWarning(
-                "The number of output features of mod1_layers and mod2_layers must be the same for Activation fusion. Please change the final layers in the modality layers to have the same number of output features as each other."
+                "The number of output features of the modality layers must be the same for Activation fusion. Please change the final layers in the modality layers to have the same number of output features as each other."
             )
+
+        # And if there are 3 modalities, check the third one
+        if self.data_dims["mod3_dim"] is not None:
+            check_model_validity.check_dtype(
+                self.mod3_layers, nn.ModuleDict, "mod3_layers"
+            )
[…]
