### 2024-04-22 4ee1b1c
Altered attention and activation to be able to take three modalities and also made it user-modifiable of which modality has the self attention
