+++ tests/test_models/test_subspace_and_graph_methods.py
@@ -6,6 +6,7 @@ import pandas as pd
 from unittest.mock import patch, Mock
 from unittest import mock
 import lightning.pytorch as pl
+
 # from lightning.pytorch.callbacks import EarlyStopping
 from torch import nn
 from torch_geometric.data import Data
@@ -16,19 +17,23 @@ from fusilli.data import CustomDataset
 from fusilli.fusionmodels.tabularimagefusion.denoise_tab_img_maps import (
     DenoisingAutoencoder,
     ImgUnimodalDAE,
-    denoising_autoencoder_subspace_method
+    denoising_autoencoder_subspace_method,
 )
 from fusilli.fusionmodels.tabularfusion.edge_corr_gnn import EdgeCorrGraphMaker
 
 from ..test_data.test_TrainTestDataModule import create_test_files
 from fusilli.data import TrainTestDataModule
+
 # from fusilli.utils.mcvae.src.mcvae.models import Mcvae
 # from fusilli.utils.training_utils import init_trainer
 from fusilli.fusionmodels.tabularimagefusion.concat_img_latent_tab_doubletrain import (
-    concat_img_latent_tab_subspace_method, ImgLatentSpace
+    concat_img_latent_tab_subspace_method,
+    ImgLatentSpace,
 )
 
-from fusilli.fusionmodels.tabularfusion.attention_weighted_GNN import AttentionWeightedGraphMaker
+from fusilli.fusionmodels.tabularfusion.attention_weighted_GNN import (
+    AttentionWeightedGraphMaker,
+)
 
 
 class MockFusionModel:
@@ -52,14 +57,15 @@ def sample_datamodule(create_test_files):
     fusion_model = MockFusionModel()
 
     # Call the prepare_fusion_data function with custom fusion type (non-graph)
-    dm = TrainTestDataModule(fusion_model=MockFusionModel,
-                             sources=[tabular1_csv, tabular2_csv, image_torch_file_2d],
-                             output_paths=None,
-                             prediction_task="binary",
-                             batch_size=8,
-                             test_size=0.3,
-                             multiclass_dimensions=None,
-                             )
+    dm = TrainTestDataModule(
+        fusion_model=MockFusionModel,
+        sources=[tabular1_csv, tabular2_csv, image_torch_file_2d],
+        output_paths=None,
+        prediction_task="binary",
+        batch_size=8,
+        test_size=0.3,
+        multiclass_dimensions=None,
+    )
 
     # dm = prepare_fusion_data(fusion_model, params)
     dm.prepare_data()
@@ -71,6 +77,7 @@ def sample_datamodule(create_test_files):
 class MockFusionTabImgModel:
     fusion_type = "subspace"
     modality_type = "tabular_image"
+    three_modalities = False
 
     def __init__(self):
         pass
@@ -89,14 +96,19 @@ def sample_tabimg_datamodule(create_test_files):
     fusion_model = MockFusionTabImgModel()
 
     # Call the prepare_fusion_data function with custom fusion type (non-graph)
-    dm = TrainTestDataModule(fusion_model=MockFusionTabImgModel,
-                             sources=[tabular1_csv, tabular2_csv, image_torch_file_2d],
-                             output_paths=None,
-                             prediction_task="binary",
-                             batch_size=8,
-                             test_size=0.3,
-                             multiclass_dimensions=None,
-                             )
[…]
