### 2024-04-23 b90bf83
fixed test_modification tests to go along with new data dims and sources loading as dictionaries
