+++ tests/test_modifications/test_graph_modifications.py
@@ -7,7 +7,7 @@ from fusilli.fusionmodels.tabularfusion.edge_corr_gnn import (
 )
 from fusilli.fusionmodels.tabularfusion.attention_weighted_GNN import (
     AttentionWeightedGNN,
-    AttentionWeightedGraphMaker
+    AttentionWeightedGraphMaker,
 )
 from tests.test_data.test_TrainTestDataModule import create_test_files
 from fusilli.data import TrainTestGraphDataModule
@@ -19,7 +19,12 @@ from lightning.pytorch.callbacks import EarlyStopping
 @pytest.fixture
 def model_instance_EdgeCorrGNN():
     prediction_task = "regression"
-    data_dims = [10, 15, None]
+    data_dims = {
+        "mod1_dim": 10,
+        "mod2_dim": 15,
+        "mod3_dim": None,
+        "img_dim": None,
+    }
     multiclass_dimensions = None
     return EdgeCorrGNN(prediction_task, data_dims, multiclass_dimensions)
 
@@ -30,16 +35,17 @@ def model_instance_EdgeCorrGraphMaker(create_test_files):
     tabular2_csv = create_test_files["tabular2_csv"]
     image_torch_file_2d = create_test_files["image_torch_file_2d"]
 
-    sources = [tabular1_csv, tabular2_csv, image_torch_file_2d]
+    sources = {
+        "tabular1": tabular1_csv,
+        "tabular2": tabular2_csv,
+        "image": image_torch_file_2d,
+    }
     example_fusion_model = Mock()
     example_fusion_model.modality_type = "tabular_tabular"
 
     # Initialize the TrainTestDataModule
     dm = TrainTestGraphDataModule(
-        example_fusion_model,
-        sources,
-        EdgeCorrGraphMaker,
-        test_size=0.3
+        example_fusion_model, sources, EdgeCorrGraphMaker, test_size=0.3
     )
     dm.prepare_data()
 
@@ -49,7 +55,12 @@ def model_instance_EdgeCorrGraphMaker(create_test_files):
 @pytest.fixture
 def model_instance_AttentionWeightedGNN():
     prediction_task = "regression"
-    data_dims = [10, 15, None]
+    data_dims = {
+        "mod1_dim": 10,
+        "mod2_dim": 15,
+        "mod3_dim": None,
+        "img_dim": None,
+    }
     multiclass_dimensions = None
     return AttentionWeightedGNN(prediction_task, data_dims, multiclass_dimensions)
 
@@ -64,16 +75,17 @@ def model_instance_AttentionWeightedGraphMaker(create_test_files):
         "test_size": 0.2,
     }
 
-    sources = [tabular1_csv, tabular2_csv, image_torch_file_2d]
+    sources = {
+        "tabular1": tabular1_csv,
+        "tabular2": tabular2_csv,
+        "image": image_torch_file_2d,
+    }
     example_fusion_model = Mock()
     example_fusion_model.modality_type = "tabular_tabular"
 
     # Initialize the TrainTestDataModule
     dm = TrainTestGraphDataModule(
-        example_fusion_model,
-        sources,
-        AttentionWeightedGraphMaker,
-        test_size=0.3
+        example_fusion_model, sources, AttentionWeightedGraphMaker, test_size=0.3
     )
     dm.prepare_data()
 
@@ -98,24 +110,22 @@ correct_modifications = {
         ),
         "dropout_prob": 0.4,
     },
-    "AttentionWeightedGraphMaker": {"early_stop_callback":
-                                        EarlyStopping(monitor="val_loss", ),
-                                    "edge_probability_threshold": 80,
-                                    "attention_MLP_test_size": 0.3,
-                                    "AttentionWeightingMLPInstance.weighting_layers": nn.ModuleDict(
-                                        {
[…]
