+++ tests/test_modifications/test_unimodal_modifications.py
@@ -9,33 +9,49 @@ from fusilli.fusionmodels.unimodal.tabular2 import Tabular2Unimodal
 @pytest.fixture
 def model_instance_ImgUnimodal_3D():
     pred_type = "regression"
-    data_dims = [None, None, [100, 100, 100]]
-    params = {}
-    return ImgUnimodal(pred_type, data_dims, params)
+    data_dims = {
+        "mod1_dim": None,
+        "mod2_dim": None,
+        "mod3_dim": None,
+        "img_dim": [100, 100, 100],
+    }
+    return ImgUnimodal(pred_type, data_dims, None)
 
 
 @pytest.fixture
 def model_instance_ImgUnimodal_2D():
     pred_type = "regression"
-    data_dims = [None, None, [100, 100]]
-    params = {}
-    return ImgUnimodal(pred_type, data_dims, params)
+    data_dims = {
+        "mod1_dim": None,
+        "mod2_dim": None,
+        "mod3_dim": None,
+        "img_dim": [100, 100],
+    }
+    return ImgUnimodal(pred_type, data_dims, None)
 
 
 @pytest.fixture
 def model_instance_Tabular1Unimodal():
     pred_type = "regression"
-    data_dims = [10, None, None]
-    params = {}
-    return Tabular1Unimodal(pred_type, data_dims, params)
+    data_dims = {
+        "mod1_dim": 10,
+        "mod2_dim": None,
+        "mod3_dim": None,
+        "img_dim": None,
+    }
+    return Tabular1Unimodal(pred_type, data_dims, None)
 
 
 @pytest.fixture
 def model_instance_Tabular2Unimodal():
     pred_type = "regression"
-    data_dims = [None, 15, None]
-    params = {}
-    return Tabular2Unimodal(pred_type, data_dims, params)
+    data_dims = {
+        "mod1_dim": None,
+        "mod2_dim": 15,
+        "mod3_dim": None,
+        "img_dim": None,
+    }
+    return Tabular2Unimodal(pred_type, data_dims, None)
 
 
 correct_modifications_2D = {
@@ -248,8 +264,8 @@ def test_correct_modify_model_architecture(model_name, model_fixture, request):
     for key, modification in correct_modifications.get(model_name, {}).items():
         if hasattr(getattr(modified_model, key), "__code__"):
             assert (
-                    getattr(modified_model, key).__code__.co_code
-                    == modification.__code__.co_code
+                getattr(modified_model, key).__code__.co_code
+                == modification.__code__.co_code
             )
         else:
             # "modification" may have been modified (ironically) by the calc_fused_layers method
@@ -263,15 +279,15 @@ def test_correct_modify_model_architecture(model_name, model_fixture, request):
     # Ensure that the final prediction layer has been modified as expected but the output dim
     # has not
     assert (
-            modified_model.final_prediction[-1].out_features
-            == original_model.final_prediction[-1].out_features
+        modified_model.final_prediction[-1].out_features
+        == original_model.final_prediction[-1].out_features
     )
 
 
 # Test the modify_model_architecture function with incorrect data type for the modifications
 @pytest.mark.parametrize("model_name, model_fixture", model_instances)
 def test_wrong_data_type_modify_model_architecture_training(
-        model_name, model_fixture, request
+    model_name, model_fixture, request
 ):
     # iterate through the modifications to check each throws an error
     for key, modification in incorrect_dtype_modifications.get(model_name, {}).items():
@@ -279,7 +295,7 @@ def test_wrong_data_type_modify_model_architecture_training(
         # Modify the model's architecture using the function
 
         with pytest.raises(
-                TypeError, match="Incorrect data type for the modifications"
[…]

[…]
