### 2024-04-23 97c6b7a
fixed tests to handle new data dims and sources loading and 3 tabular modalities
