### 2024-04-26 75d450d
fixed a suspected bug in the implementation of own early stopping callback
