### 2024-05-13 c8ac99a
changed layer mods modification to include mcvae early stopping
