### 2024-05-14 e4cc450
making tabular fusilli methods more gpu friendly - altering some layers in channelwise attention and making metric calculations device alterable
