+++ fusilli/utils/metrics_utils.py
@@ -10,16 +10,19 @@ class MetricsCalculator:
     Calculates metrics of the models and houses list of the available metrics to use.
     """
 
-    def __init__(self, base_model_instance):
+    def __init__(self, base_model_instance, device=None):
         """
 
         Parameters
         ----------
         base_model_instance : fusilli.fusionmodels.base_model.BaseModel
             Instance of the base model. Has information on the prediction task and multiclass dimensions if applicable.
+        device : torch.device
+            Device to run the model on.
         """
         self.model = base_model_instance
         self.prediction_task = base_model_instance.model.prediction_task
+
 
     def auroc(self, preds, labels, logits):
         """
@@ -41,9 +44,9 @@ class MetricsCalculator:
         """
 
         if self.prediction_task == "binary":
-            auroc_equation = tm.AUROC(task="binary")
+            auroc_equation = tm.AUROC(task="binary").to(preds.device)
         elif self.prediction_task == "multiclass":
-            auroc_equation = tm.AUROC(num_classes=self.model.multiclass_dimensions, task="multiclass")
+            auroc_equation = tm.AUROC(num_classes=self.model.multiclass_dimensions, task="multiclass").to(preds.device)
         else:
             raise ValueError("Invalid prediction task for AUROC.")
 
@@ -71,11 +74,11 @@ class MetricsCalculator:
 
         if self.prediction_task == "binary":
             # do binary accuracy
-            accuracy_equation = tm.Accuracy(task="binary")
+            accuracy_equation = tm.Accuracy(task="binary").to(preds.device)
 
         elif self.prediction_task == "multiclass":
             # do multiclass accuracy
-            accuracy_equation = tm.Accuracy(num_classes=self.model.multiclass_dimensions, task="multiclass", top_k=1)
+            accuracy_equation = tm.Accuracy(num_classes=self.model.multiclass_dimensions, task="multiclass", top_k=1).to(preds.device)
         else:
             raise ValueError("Invalid prediction task for accuracy.")
 
@@ -103,8 +106,9 @@ class MetricsCalculator:
 
         if self.prediction_task != "regression":
             raise ValueError("Invalid prediction task for R2.")
+        
 
-        return tm.R2Score()(preds, labels)
+        return tm.R2Score().to(preds.device)(preds, labels)
 
     def mse(self, preds, labels, logits):
         """
@@ -128,7 +132,7 @@ class MetricsCalculator:
         if self.prediction_task != "regression":
             raise ValueError("Invalid prediction task for mse.")
 
-        return tm.MeanSquaredError()(preds, labels)
+        return tm.MeanSquaredError().to(preds.device)(preds, labels)
 
     def mae(self, preds, labels, logits):
         """
@@ -153,7 +157,7 @@ class MetricsCalculator:
         if self.prediction_task != "regression":
             raise ValueError("Invalid prediction task for mae.")
 
-        return tm.MeanAbsoluteError()(preds, labels)
+        return tm.MeanAbsoluteError().to(preds.device)(preds, labels)
 
     def recall(self, preds, labels, logits):
         """
@@ -176,9 +180,9 @@ class MetricsCalculator:
         """
 
         if self.prediction_task == "binary":
-            recall_equation = tm.Recall(task="binary")
+            recall_equation = tm.Recall(task="binary").to(preds.device)
         elif self.prediction_task == "multiclass":
-            recall_equation = tm.Recall(num_classes=self.model.multiclass_dimensions, task="multiclass")
+            recall_equation = tm.Recall(num_classes=self.model.multiclass_dimensions, task="multiclass").to(preds.device)
         else:
             raise ValueError("Invalid prediction task for recall.")
 
@@ -205,9 +209,9 @@ class MetricsCalculator:
         """
 
[…]
