### 2024-05-15 7c814e5
added training_modifications argument to train_and_save_models where the accelerator and number of devices can be specified. Also changed the metrics_utils calculations to put the metric equations onto whatever device the data is on.
