### 2024-05-15 84af094
added documentation for GPU on customising training section. Currently not implemented for subspace methods but that is on the list
