### 2024-05-15 efe143d
Added the early stopping patience and tolerance to the layer mods dictionary options so it is more easily customisable
