### 2024-05-15 6952440
Fixed tests in test_pipeline to remove params arguments
