### 2024-05-15 93d4a14
Fixed bug where mcvae patience for early stopping was called self.patience and self.mcvae_patience
