### 2024-05-15 6f1baab
fixing early stopping test
