### 2024-05-16 2f46649
Added training modifications dict as an argument to prepare_fusion_data instead of only train and save models, with the plan to remove it from train altogether and just use the datamodule attribute version
