+++ fusilli/data.py
@@ -456,6 +456,7 @@ class TrainTestDataModule(pl.LightningDataModule):
         own_early_stopping_callback=None,
         num_workers=0,
         test_indices=None,
+        training_modifications=None,
         kwargs=None,
     ):
         """
@@ -496,6 +497,9 @@ class TrainTestDataModule(pl.LightningDataModule):
         test_indices : list
             List of indices to use for testing (default None). If None, the test indices are
             randomly selected using the test_size parameter.
+        training_modifications : dict
+            Dictionary of training modifications to make to the subspace method.
+            Keys are "accelerator" and "devices"
         kwargs : dict
             Dictionary of extra arguments for the subspace method class.
         """
@@ -530,6 +534,7 @@ class TrainTestDataModule(pl.LightningDataModule):
         self.own_early_stopping_callback = own_early_stopping_callback
         self.num_workers = num_workers
         self.test_indices = test_indices
+        self.training_modifications = training_modifications
         self.kwargs = kwargs
 
     def prepare_data(self):
@@ -591,6 +596,7 @@ class TrainTestDataModule(pl.LightningDataModule):
                     max_epochs=self.max_epochs,
                     k=None,
                     train_subspace=True,
+                    training_modifications=self.training_modifications,
                 )
 
                 # modify the subspace method architecture if specified
@@ -623,7 +629,7 @@ class TrainTestDataModule(pl.LightningDataModule):
                 # we have already trained the subspace method, so load it from the checkpoint
 
                 self.subspace_method_train = self.subspace_method(
-                    self, max_epochs=self.max_epochs, k=None, train_subspace=False
+                    self, max_epochs=self.max_epochs, k=None, train_subspace=False, training_modifications=self.training_modifications
                 )  # will return a init subspace method with the subspace models as instance attributes
 
                 # modify the subspace method architecture if specified
@@ -739,6 +745,9 @@ class KFoldDataModule(pl.LightningDataModule):
         List of indices to use for k-fold cross validation (default None). If None, the k-fold
         indices are randomly selected. Structure is a list of tuples of (train_indices,
         test_indices). Must be the same length as num_folds.
+    trainings_modifications : dict
+        Dictionary of training modifications to make to the subspace method.
+        Keys are "accelerator" and "devices"
     kwargs : dict
         Dictionary of extra arguments for the subspace method class.
     """
@@ -760,6 +769,7 @@ class KFoldDataModule(pl.LightningDataModule):
         own_early_stopping_callback=None,
         num_workers=0,
         own_kfold_indices=None,
+        training_modifications=None,
         kwargs=None,
     ):
         """
@@ -803,6 +813,9 @@ class KFoldDataModule(pl.LightningDataModule):
             List of indices to use for k-fold cross validation (default None). If None, the k-fold
             indices are randomly selected. Structure is a list of tuples of (train_indices,
             test_indices). Must be the same length as num_folds.
+        training_modifications : dict
+            Dictionary of training modifications to make to the subspace method.
+            Keys are "accelerator" and "devices"
         kwargs : dict
             Dictionary of extra arguments for the subspace method class.
         """
@@ -844,6 +857,7 @@ class KFoldDataModule(pl.LightningDataModule):
         self.own_early_stopping_callback = own_early_stopping_callback
         self.num_workers = num_workers
         self.own_kfold_indices = own_kfold_indices
+        self.training_modifications = training_modifications
         self.kwargs = kwargs
 
     def prepare_data(self):
@@ -929,6 +943,7 @@ class KFoldDataModule(pl.LightningDataModule):
                         k=k,
                         max_epochs=self.max_epochs,
                         train_subspace=True,
+                        training_modifications=self.training_modifications,
[…]
