+++ fusilli/eval.py
@@ -180,6 +180,7 @@ class ParentPlotter:
         test_data_paths,
         checkpoint_file_suffix=None,
         layer_mods=None,
+        training_modifications=None,
     ):
         """
         Get new data by running through trained model for a kfold model.
@@ -198,6 +199,8 @@ class ParentPlotter:
             Default is None.
         layer_mods: dict, optional
             Dictionary of the layer modifications to make to the model.
+        training_modifications: dict, optional
+            Dictionary of the training modifications to make to the model.
 
         Returns
         -------
@@ -288,6 +291,7 @@ class ParentPlotter:
                 num_folds=num_folds,
                 checkpoint_path=subspace_ckpts,
                 layer_mods=layer_mods,
+                training_modifications=training_modifications,
             )
 
             # just taking the first fold because we don't need to split the new data into folds
@@ -400,6 +404,7 @@ class ParentPlotter:
         test_data_paths,
         checkpoint_file_suffix=None,
         layer_mods=None,
+        training_modifications=None,
     ):
         """
         Get new data by running through trained model for a train/test model.
@@ -417,6 +422,8 @@ class ParentPlotter:
             Default is None.
         layer_mods: dict, optional
             Dictionary of the layer modifications to make to the model.
+        training_modifications: dict, optional
+            Dictionary of the training modifications to make to the model.
 
         Returns
         -------
@@ -448,6 +455,7 @@ class ParentPlotter:
         # ckpt_path = model[0][1]
         model = model_list[0]
 
+
         model.eval()
 
         if hasattr(model.model, "graph_maker"):
@@ -456,6 +464,7 @@ class ParentPlotter:
             )
 
         if model.model.subspace_method is not None:
+            
             subspace_ckpts = []
             for subspace_model in model.model.subspace_method.subspace_models:
                 subspace_ckpts.append(
@@ -480,6 +489,7 @@ class ParentPlotter:
             output_paths=output_paths,
             checkpoint_path=subspace_ckpts,
             layer_mods=layer_mods,
+            training_modifications=training_modifications,
         )
 
         # concatenating the train and test datasets because we want to get the predictions for all the data
@@ -521,8 +531,12 @@ class ParentPlotter:
         logits_list = []
         reals_list = []
 
+
         for batch in dataloader:
             x, y = new_model.get_data_from_batch(batch)
+            # if new_model.model.subspace_method is not None:
+            #     out = new_model.get_model_outputs_and_loss(x.cpu().detach(), y.cpu().detach())
+            # else:
             out = new_model.get_model_outputs_and_loss(x, y)
             loss, end_output, logits = out
 
@@ -569,6 +583,7 @@ class RealsVsPreds(ParentPlotter):
         test_data_paths,
         checkpoint_file_suffix=None,
         layer_mods=None,
+        training_modifications=None
     ):
         """
 
@@ -629,6 +644,7 @@ class RealsVsPreds(ParentPlotter):
                 test_data_paths,
                 checkpoint_file_suffix,
                 layer_mods,
+                training_modifications
             )
 
             figure = cls.reals_vs_preds_kfold(
@@ -655,6 +671,7 @@ class RealsVsPreds(ParentPlotter):
                 test_data_paths,
                 checkpoint_file_suffix,
                 layer_mods,
+                training_modifications
             )
 
             # plot the figure
@@ -971,6 +988,7 @@ class ConfusionMatrix(ParentPlotter):
         test_data_paths,
         checkpoint_file_suffix=None,
         layer_mods=None,
+        training_modifications=None
     ):
         """
         Confusion matrix using new data (i.e. data that was not used to train or validate the model).
@@ -987,6 +1005,12 @@ class ConfusionMatrix(ParentPlotter):
[…]
