### 2024-06-19 #6414 @florencejt
Hi @atrisovic! Thanks for your help on this, it's a shame about the other reviewers. I've got two suggestions for new potential reviewers: Sam Gijsen (sam.gijsen@charite.de) and Cameron Shand (cameron.shand@crick.ac.uk). Sam has used fusilli before so is acquainted with how it works, and Cameron is a research software engineer.
Let me know if these suggestions are ok and if you'd like me to try to think of more.
