### 2024-08-28 a60da0d
adding shap guidance work in progress docs
