+++ docs/examples/customising_behaviour/plot_add_explainability.py
@@ -1,0 +1,18 @@
+"""
+Adding Explainability to Fusilli
+################################################
+
+Adding SHAP values using the shap library to fusilli.
+Guidance includes for :
+- Unimodal models
+    - Tabular models
+    - Image models (2D only)
+- Multimodal models (tabular and tabular)
+- Multimodal models (tabular and image)
+
+.. warning::
+
+    Disclaimer: I'm not an expert on SHAP values, I just recently starting using them with this library so hopefully this can help somebody else to get started.
+
+
+"""
