### 2024-08-28 7222346
Added argument to the from_new_data eval functions to input the multiclass dimensions if using multiclass classification
