### 2024-09-05 8c47174
un-tensor-fied outputs from eval functions
