### 2024-09-12 6264087
added checks to see if csv has unnamed : 0 in columns and also made it so wandb uses the environemnt variable for where it stores its logs locally
