+++ fusilli/utils/training_utils.py
@@ -114,7 +114,8 @@ def set_logger(
             project_name = "fusilli"
 
         logger = WandbLogger(
-            save_dir=os.getcwd() + "/logs",
+            # save_dir=os.getcwd() + "/logs",
+            save_dir=os.environ["WANDB_DIR"],
             project=project_name,
             name=name,
             tags=tags,
