### 2024-09-18 bbd29f9
Adding external test into the title of from_new_data plots
