+++ fusilli/eval.py
@@ -110,7 +110,9 @@ class ParentPlotter:
                 logits=model_list[0].safe_squeeze(all_val_logits),
             )
 
-            overall_kfold_metrics[metric_name.lower()] = val_metric.cpu().detach().item()
+            overall_kfold_metrics[metric_name.lower()] = (
+                val_metric.cpu().detach().item()
+            )
 
         return (
             train_reals,
@@ -172,7 +174,12 @@ class ParentPlotter:
 
     @classmethod
     def get_new_kfold_data(
-            cls, model_list, output_paths, test_data_paths, checkpoint_file_suffix=None, layer_mods=None
+        cls,
+        model_list,
+        output_paths,
+        test_data_paths,
+        checkpoint_file_suffix=None,
+        layer_mods=None,
     ):
         """
         Get new data by running through trained model for a kfold model.
@@ -296,7 +303,7 @@ class ParentPlotter:
                     checkpoint_dir=output_paths["checkpoints"],
                     model=model,
                     checkpoint_file_suffix=checkpoint_file_suffix,
-                    fold=k
+                    fold=k,
                 )
             )
 
@@ -317,7 +324,9 @@ class ParentPlotter:
                     layer_mods,
                 )
             # load the state dict
-            new_model.load_state_dict(torch.load(trained_fusion_model_checkpoint)["state_dict"])
+            new_model.load_state_dict(
+                torch.load(trained_fusion_model_checkpoint)["state_dict"]
+            )
 
             new_model.eval()
 
@@ -370,7 +379,9 @@ class ParentPlotter:
                 logits=model_list[0].safe_squeeze(all_val_logits),
             )
 
-            overall_kfold_metrics[metric_name.lower()] = val_metric.cpu().detach().item()
+            overall_kfold_metrics[metric_name.lower()] = (
+                val_metric.cpu().detach().item()
+            )
 
         return (
             train_reals,
@@ -383,7 +394,12 @@ class ParentPlotter:
 
     @classmethod
     def get_new_tt_data(
-            cls, model_list, output_paths, test_data_paths, checkpoint_file_suffix=None, layer_mods=None
+        cls,
+        model_list,
+        output_paths,
+        test_data_paths,
+        checkpoint_file_suffix=None,
+        layer_mods=None,
     ):
         """
         Get new data by running through trained model for a train/test model.
@@ -495,7 +511,9 @@ class ParentPlotter:
                 layer_mods,
             )
         # load the state dict
-        new_model.load_state_dict(torch.load(trained_fusion_model_checkpoint)["state_dict"])
+        new_model.load_state_dict(
+            torch.load(trained_fusion_model_checkpoint)["state_dict"]
+        )
 
         new_model.eval()
         # get the predictions
@@ -545,7 +563,12 @@ class RealsVsPreds(ParentPlotter):
 
     @classmethod
     def from_new_data(
-            cls, model_list, output_paths, test_data_paths, checkpoint_file_suffix=None, layer_mods=None
+        cls,
+        model_list,
+        output_paths,
+        test_data_paths,
+        checkpoint_file_suffix=None,
+        layer_mods=None,
     ):
         """
 
@@ -601,7 +624,11 @@ class RealsVsPreds(ParentPlotter):
                 metrics_per_fold,
                 overall_kfold_metrics,
             ) = cls.get_new_kfold_data(
-                model_list, output_paths, test_data_paths, checkpoint_file_suffix, layer_mods
+                model_list,
+                output_paths,
+                test_data_paths,
+                checkpoint_file_suffix,
+                layer_mods,
             )
 
             figure = cls.reals_vs_preds_kfold(
@@ -623,7 +650,11 @@ class RealsVsPreds(ParentPlotter):
                 val_preds,
                 metric_values,
             ) = cls.get_new_tt_data(
[…]
