### 2024-09-20 6277dc8
trying to find fix to when img dims are too small for the dummy conv output - need to commit to get it on the cluster
