+++ fusilli/fusionmodels/unimodal/image.py
@@ -65,12 +65,13 @@ class ImgUnimodal(ParentFusionModel, nn.Module):
         -------
         None
         """
-
-        dummy_conv_output = Variable(
-            torch.rand((1,) + tuple(self.data_dims["img_dim"]))
-        )
-        for layer in self.img_layers.values():
-            dummy_conv_output = layer(dummy_conv_output)
+        print(self.data_dims["img_dim"])
+        try:
+            dummy_conv_output = Variable(
+                torch.rand((1,) + tuple(self.data_dims["img_dim"]))
+            )
+            for layer in self.img_layers.values():
+                dummy_conv_output = layer(dummy_conv_output)
 
         flattened_img_output_size = dummy_conv_output.data.view(1, -1).size(1)
 
