### 2024-10-10 a8605e7
making reals and preds returned from model comparison for my own project
