+++ fusilli/fusionmodels/unimodal/image.py
@@ -65,13 +65,15 @@ class ImgUnimodal(ParentFusionModel, nn.Module):
         -------
         None
         """
-        print(self.data_dims["img_dim"])
+        # print(self.data_dims["img_dim"])
+        dummy_conv_output = Variable(
+            torch.rand((1,) + tuple(self.data_dims["img_dim"]))
+        )
         try:
-            dummy_conv_output = Variable(
-                torch.rand((1,) + tuple(self.data_dims["img_dim"]))
-            )
             for layer in self.img_layers.values():
                 dummy_conv_output = layer(dummy_conv_output)
+        except:
+            pass
 
         flattened_img_output_size = dummy_conv_output.data.view(1, -1).size(1)
 
