### 2024-10-10 ec6ad58
making image methods not throw an error if image input is too small for them to pass through the default network size. it won't work if not using layer mods, but if the layer mods make it the correct network size, itll be fine
