+++ fusilli/fusionmodels/tabularimagefusion/concat_img_maps_tabular_maps.py
@@ -79,11 +79,14 @@ class ConcatImageMapsTabularMaps(ParentFusionModel, nn.Module):
         None
         """
         # get flattened image output size
-        dummy_conv_output = Variable(
-            torch.rand((1,) + tuple(self.data_dims["img_dim"]))
-        )
-        for layer in self.img_layers.values():
-            dummy_conv_output = layer(dummy_conv_output)
+        try:
+            dummy_conv_output = Variable(
+                torch.rand((1,) + tuple(self.data_dims["img_dim"]))
+            )
+            for layer in self.img_layers.values():
+                dummy_conv_output = layer(dummy_conv_output)
+        except:
+            pass
         flattened_img_size = dummy_conv_output.data.view(1, -1).size(1)
 
         # get tabular output size
