+++ fusilli/fusionmodels/unimodal/image.py
@@ -65,11 +65,11 @@ class ImgUnimodal(ParentFusionModel, nn.Module):
         -------
         None
         """
-        print(self.data_dims["img_dim"])
+        # print(self.data_dims["img_dim"])
+        dummy_conv_output = Variable(
+            torch.rand((1,) + tuple(self.data_dims["img_dim"]))
+        )
         try:
-            dummy_conv_output = Variable(
-                torch.rand((1,) + tuple(self.data_dims["img_dim"]))
-            )
             for layer in self.img_layers.values():
                 dummy_conv_output = layer(dummy_conv_output)
         except:
