### 2024-10-17 0ba2f51
adding check that will make attention reduction ratio 2 by default if its too big for data
