### 2025-02-03 3062e2a
adding image downsample size to eval from new data functions
