### 2025-02-05 e8cf121
removed training modifications argument in training subspace method because i hadn't implemented it yet
