### 2025-02-07 09f0eeb
better checking of integer latent dims
