### 2025-02-28 3c97190
adding ability to enter augmentations for imaging
