+++ fusilli/data.py
@@ -15,6 +15,8 @@ from torch.utils.data import DataLoader, Dataset
 from torch_geometric.data.lightning import LightningNodeData
 from fusilli.utils import model_modifier
 import warnings
+import matplotlib.pyplot as plt
+import os
 
 
 def check_for_unnamed_index_column(df):
@@ -131,10 +133,11 @@ class CustomDataset(Dataset):
         Tensor of predictive features for uni-modal data.
     labels : tensor
         Tensor of labels.
-
+    transforms : list or None
+        List of transforms to apply on the image. Will only be used for image-only (unimodal) data.
     """
 
-    def __init__(self, pred_features, labels):
+    def __init__(self, pred_features, labels, transforms=None):
         """
         Parameters
         ----------
@@ -144,6 +147,9 @@ class CustomDataset(Dataset):
             (i.e. tabular or image data without labels).
         labels : dataframe
             Dataframe of labels (column name must be "prediction_label").
+        transforms : list
+            List of transforms to apply to the images. Such as torchio.transforms functions with the corresponding arguments filled in.
+            (default None)
 
         Raises
         ------
@@ -185,6 +191,8 @@ class CustomDataset(Dataset):
         else:
             self.labels = self.labels.float()
 
+        self.transforms = transforms
+
     def __len__(self):
         """
         Returns the length of the dataset.
@@ -208,8 +216,22 @@ class CustomDataset(Dataset):
         idx : int
             Index of the item to return.
         """
+        # if self.multimodal_flag:
+        #     if self.three_modalities:
+        #         return (
+        #             self.dataset1[idx],
+        #             self.dataset2[idx],
+        #             self.dataset3[idx],
+        #             self.labels[idx],
+        #         )
+        #     else:
+        #         return self.dataset1[idx], self.dataset2[idx], self.labels[idx]
+        # else:
+        #     return self.dataset[idx], self.labels[idx]
+
         if self.multimodal_flag:
             if self.three_modalities:
+                # not doing any transforms for 3-modalities (all tabular)
                 return (
                     self.dataset1[idx],
                     self.dataset2[idx],
@@ -217,9 +239,56 @@ class CustomDataset(Dataset):
                     self.labels[idx],
                 )
             else:
-                return self.dataset1[idx], self.dataset2[idx], self.labels[idx]
+                # For 2-modality multimodal data, assume modality 2 is the image.
+                tabular = self.dataset1[idx]
+
+                image = self.dataset2[idx]
+
+                # Apply transforms to the image if provided.
+                # Transforms will only not be None when called from image-only or tabular-image dataloaders.
+                if self.transforms is not None:
+
+                    # self.save_middle_slice(image, idx, non_augment=True)
+
+                    for t in self.transforms:
+                        print("Transform being applied:", t)
+                        print("Image shape:", image.shape)
+                        image = t(image)
+
+                    # self.save_middle_slice(image, idx)
+
+                return tabular, image, self.labels[idx]
         else:
-            return self.dataset[idx], self.labels[idx]
+            # For unimodal (image-only) data.
+            sample = self.dataset[idx]
+            # Transforms will only not be None when called from image-only or tabular-image dataloaders.
+            if self.transforms is not None:
+                for t in self.transforms:
+                    print("Transform being applied:", t)
+                    sample = t(sample)
+
+            return sample, self.labels[idx]
+
+    def save_middle_slice(self, image, idx, non_augment=False):
+        """Saves the middle slice of a 3D image tensor as a PNG for debugging."""
[…]
