### 2025-03-25 3a2a010
making transforms work with 2d images
