### 2025-03-26 21dbb32
getting logits out of models
