### 2025-03-28 9719733
removing print statements from concatimagemapstabularmaps
