### 2025-03-28 7e2167d
improving printing in eval
