### 2025-07-02 1fa5359
adding train logits recording
