### 2025-07-07 944fb04
adding multiclass dimensions argument to confusion matrix kfold from new data
