### 2025-07-23 c745aaa
investigating image losses
