### 2025-07-23 ca156b2
investigating image losses
