### 2025-07-23 2fd3b77
fixed binary activation function
