### 2026-02-04 #6414 @atrisovic
Hi all,

Following up — I agree that if @florencejt would like to proceed with the submission, we should look for a new set of reviewers.

@florencejt, please let us know this week whether you’d like to continue with the JOSS review.

Best,
Ana
