### 2026-02-07 #6414 @florencejt
Hi @atrisovic and @crvernon

Sorry for the slow response. I definitely still want to proceed with the review for Fusilli if possible.

The main reason the code hasn't been updated on GitHub is that I wanted to keep the codebase stable while the review was pending. As the process ended up taking quite a while, I stayed in a bit of a holding pattern with the main branch to make sure everything stayed consistent for the reviewers. In the meantime, I’ve started a full time job and I'm still finishing off my thesis, so it’s been a bit of a busy period!

I've been using this code for my PhD research and have added a few new features locally, but the submitted paper currently reflects the version on GitHub. Would you prefer me to keep it as it is so it matches the paper, or should I push those minor updates before the new reviewers start?

I’m happy to go with whatever works best for you, and I’m fine with starting fresh with new reviewers.

Best wishes,
Florence
