### 2026-02-09 #6414 @cairola
Hi @atrisovic <https://github.com/atrisovic> and @crvernon
<https://github.com/crvernon>,

Apologies for not taking action on this task up until now. I am available
to review the paper and codebase if you still need it.
Maybe, can you briefly explain me the process to follow? I recall I
generated the checklist, but was not clear how to actually go through it.

Thank you.

Best Regards,
Andrea

On Sat, Feb 7, 2026 at 9:43 AM Florence Townend ***@***.***>
wrote:

> *florencejt* left a comment (openjournals/joss-reviews#6414)
> <https://github.com/openjournals/joss-reviews/issues/6414#issuecomment-3864036951>
>
> Hi @atrisovic <https://github.com/atrisovic> and @crvernon
> <https://github.com/crvernon>
>
> Sorry for the slow response. I definitely still want to proceed with the
> review for Fusilli if possible.
>
> The main reason the code hasn't been updated on GitHub is that I wanted to
> keep the codebase stable while the review was pending. As the process ended
> up taking quite a while, I stayed in a bit of a holding pattern with the
> main branch to make sure everything stayed consistent for the reviewers. In
> the meantime, I’ve started a full time job and I'm still finishing off my
> thesis, so it’s been a bit of a busy period!
>
> I've been using this code for my PhD research and have added a few new
> features locally, but the submitted paper currently reflects the version on
> GitHub. Would you prefer me to keep it as it is so it matches the paper, or
> should I push those minor updates before the new reviewers start?
>
> I’m happy to go with whatever works best for you, and I’m fine with
> starting fresh with new reviewers.
>
> Best wishes,
> Florence
>
