### 2026-07-29 d428cc9
Fix urllib3 CVEs and remove hardcoded wandb API key
