### 2026-07-29 7570517
Fix failing tests and bugs from leftover debug/WIP code
- restore label dtype conversion (long for classification) that was commented out
- fix stale test expectations after Sigmoid removal / output squeeze changes
- catch ValueError alongside RuntimeError when train batch lists are empty
  (standalone trainer.validate() call has no preceding training_step)
- add missing training_modifications param to two subspace method classes
- fix check_dtype() reporting wrong type name
- fix CI: checkout wasn't fetching the mcvae submodule, so mcvae_model.py had
  0% test coverage; switch back to the sub
[…]
