+++ fusilli/fusionmodels/tabularfusion/mcvae_model.py
@@ -7,10 +7,7 @@ import torch.nn as nn
 from fusilli.fusionmodels.base_model import ParentFusionModel
 import torch
 
-# from fusilli.utils.mcvae.src.mcvae.models import Mcvae
-
-# FOR LOCAL INSTALL - CHANGE WHEN PUSHING TO RELEASE
-from mcvae.models import Mcvae
+from fusilli.utils.mcvae.src.mcvae.models import Mcvae
 
 import contextlib
 import pandas as pd
