### 2026-07-29 905953a
Install wandb in CI so test_logging.py mocks can resolve wandb.init
