### 2026-07-29 98b1a44
Fix broken doc examples, stop tracking generated docs output
README and logging_with_wandb.rst had a broken import and a wrong kwarg name.
Also untracking docs/auto_examples, contributing_examples, autosummary, and
the loss_logs/figures under docs/examples - these all get regenerated on
every sphinx build so committing them just caused stale build warnings.
