+++ docs/auto_examples/creating_your_own_fusion_model/template_other_fusion.py
@@ -1,170 +1,0 @@
-"""
-General fusion template
-===============================================
-
-This is a template for creating your own fusion model: operation-based, attention-based, or tensor-based.
-If you want to implement a graph-based or subspace-based fusion model, please refer to the other templates. They require additional functions to be implemented.
-
-There are two ways to create your own fusion model:
-
-1. You can use the preset layers in the :class:`~fusilli.fusion_models.base_model.ParentFusionModel` class. This is the easiest way to create your own fusion model. You can see an example of this in the :class:`~fusilli.fusion_models.concat_tabular_data.ConcatTabularData` class.
-2. You can create your own layers. This is the most flexible way to create your own fusion model.
-
-The most important thing to remember is to output the final prediction as a list. This is because in subspace methods, the output is the label output and a reconstruction. This project's architecture is designed to be flexible enough to handle subspace methods too, hence the need for a list.
-
-Have fun creating your own fusion model!
-"""
-
-import torch.nn as nn
-import torch
-
-# importing the parent fusion model class
-from fusilli.fusionmodels.base_model import ParentFusionModel
-
-
-class TemplateFusionModel(ParentFusionModel, nn.Module):
-    """
-    Do you want to create your own multi-modal fusion model? This is the template for you!
-
-    Attributes
-    ----------
-    method_name : str
-        Name of the method.
-    modality_type : str
-        Type of modality.
-    fusion_type : str
-        Type of fusion.
-    pred_type : str
-        Type of prediction to be performed.
-    mod1_layers : dict
-        Dictionary containing the layers of the first modality.
-    mod2_layers : dict
-        Dictionary containing the layers of the second modality.
-    img_layers : dict
-        Dictionary containing the layers of the image modality.
-    fused_dim : int
-        Dimension of the fused layers.
-    fused_layers : nn.Sequential
-        Sequential layer containing the fused layers.
-    final_prediction : nn.Sequential
-        Sequential layer containing the final prediction layers. The final prediction layers
-        take in the number of features of the fused layers as input.
-
-    """
-
-    # str: name of the method
-    method_name = "Template fusion model"
-    # str: modality type
-    modality_type = "tabular_tabular"  # or "tabular1", "tabular2", "tabular_tabular", "tabular_image"
-    # str: fusion type
-    fusion_type = "attention"  # or "operation", "tensor", "graph", "subspace"
-
-    def __init__(self, pred_type, data_dims, params):
-        """
-        Initialising the model.
-
-        Parameters
-        ----------
-
-        pred_type : str
-            Type of prediction to be performed.
-        data_dims : dict
-            Dictionary containing the dimensions of the data.
-        params : dict
-            Dictionary containing the parameters of the model.
-        """
-        ParentFusionModel.__init__(self, pred_type, data_dims, params)
-        self.pred_type = pred_type
-
-        ################################
-        # SETTING THE UNI-MODAL LAYERS #
-        ################################
-        # You can either set the layers to be consistent with the rest of the library
-        # or you can set your own layers.
-
-        # USING PARENTFUSIONMODEL PRESET LAYERS
-        self.set_mod1_layers()  # set the layers for the first tabular modality
-        self.set_mod2_layers()  # set the layers for the second tabular modality
-        self.set_img_layers()  # set the layers for the image modality (if using)
-
-    def calc_fused_layers(self):
-        """
-        Calculating the fused layers.
-
-        This is here so that if mod1_layers, mod2_layers, or img_layers are changed, the fused layers are automatically recalculated
-        to make sure that there aren't dimension mismatches.
-
-        Add any errors here if your method needs specific conditions to be met.
-        For example, mod1_layers and mod2_layers must have the same number of layers.
-
[…]
