+++ docs/auto_examples/creating_your_own_fusion_model/template_graph_fusion.py
@@ -1,4 +1,0 @@
-"""
-Graph fusion template
-=====================
-"""
