+++ docs/auto_examples/creating_your_own_fusion_model/template_subspace_fusion.py
@@ -1,5 +1,0 @@
-"""
-Subspace fusion template
-========================
-
-"""

[…]
