### 2026-07-30 8bdb363
Exclude sphinx-gallery source dirs from normal doc discovery
docs/examples and docs/how_to_contribute are consumed by sphinx-gallery to
generate auto_examples/contributing_examples, but weren't in exclude_patterns,
so sphinx was also building them directly as duplicate/orphaned pages. Also
fixes a missing blank line in data_loading.rst. Cuts build warnings 34 -> 15,
remaining ones are pre-existing cosmetic issues (emoji throwing off title
underline length, one docstring cross-reference).
