### 2026-07-30 1f7a6fb
Restore sphinx-gallery cache, fixing ReadTheDocs build timeout
Untracking docs/auto_examples and docs/contributing_examples in the last
commit was wrong - sphinx_gallery_conf has run_stale_examples=False, which
relies on those committed files as a cache to skip re-running the example
scripts (they train real models) on a fresh checkout. Without them, RTD's
clean clone had to execute every example from scratch, which alone takes
14+ minutes locally and blew through RTD's 15 minute build limit.

Restored tracking for those plus the loss_logs/loss_figures side effects
under docs/examples. Rebuild
[…]
