+++ docs/contributing_examples/template_graph_fusion.py
@@ -1,0 +1,155 @@
+"""
+Fusion Model Template: Graph-based Fusion
+=========================================
+
+This template is for creating your own fusion model that is graph-based. An example of a graph-based fusion model is :class:`~fusilli.fusionmodels.tabularfusion.edge_corr_gnn.EdgeCorrGNN`.
+
+.. note::
+
+    I recommend looking at :ref:`how_to_contribute_a_template_other_fusion` before looking at this template, as I will skip over some of the details that are covered in that template (particularly regarding documentation and idiosyncrasies of the fusion model template).
+
+
+Building a graph-based fusion model is a bit different to the general template in
+:ref:`how_to_contribute_a_template_other_fusion`. The main difference is that you need to create a method that will create the graph structure from the input data.
+
+For the :class:`~fusilli.fusionmodels.tabularfusion.edge_corr_gnn.EdgeCorrGNN`, this is done in the :class:`~fusilli.fusionmodels.tabularfusion.edge_corr_gnn.EdgeCorrGraphMaker` class, which is in the same ``.py`` file as the :class:`~fusilli.fusionmodels.tabularfusion.edge_corr_gnn.EdgeCorrGNN` class.
+
+First, let's look at creating the graph-maker class.
+"""
+
+# %%
+# Creating the Graph-Maker Class
+# ------------------------------
+#
+# The graph will probably be created with the ``PyTorch Geometric`` library, which is a library for creating graph-based models in PyTorch.
+#
+# Let's import the libraries that we need:
+
+# sphinx_gallery_thumbnail_path = '_static/EdgeCorrGNN.png'
+
+import numpy as np
+import torch
+import torch.nn as nn
+
+from torch_geometric.data import Data
+
+from fusilli.fusionmodels.base_model import ParentFusionModel
+from fusilli.utils import check_model_validity
+
+
+# %%
+# Now let's create the graph-maker class.
+#
+# The graph-maker class must have the following methods:
+#
+# - ``__init__``: This method initialises the graph-maker class. It must take a ``torch.utils.data.Dataset`` as an argument (created in :meth:`.TrainTestGraphDataModule.setup` or :meth:`.KFoldGraphDataModule.setup`).
+# - ``check_params``: This method checks the parameters of the graph-maker class. It should raise a ``ValueError`` if the parameters are invalid. This will check validity of any modifications made to the model as well.
+# - ``make_graph``: This method creates the graph data structure. It must return a ``torch_geometric.data.Data`` object.
+#
+#
+
+
+class TemplateGraphMaker:
+    def __init__(self, dataset):
+        self.dataset = dataset
+
+        # other attributes for the graph maker go here
+
+    def check_params(self):
+        # check the parameters of the graph maker here
+
+        pass
+
+    def make_graph(self):
+        # create the graph here with self.dataset
+
+        self.check_params()
+
+        modality_1_data = self.dataset[:][0]
+        modality_2_data = self.dataset[:][1]
+        labels = self.dataset[:][2]
+
+        # some code to create the graph to get out:
+        # - node attributes
+        # - edge attributes
+        # - edge indices
+
+        # replace the strings with the actual graph data
+
+        data = Data(
+            x="node attributes",
+            edge_attr="edge attributes",
+            edge_index="edge indices",
+            y="labels"
+        )
+
+        return data
+
+
+# %%
+# Creating the Fusion Model Class
+# -------------------------------
+#
+# Now let's create the fusion model class that will take in the graph data structure and perform the prediction.
+#
[…]
