+++ docs/auto_examples/training_and_testing/plot_using_external_data.rst
@@ -1,0 +1,259 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "auto_examples/training_and_testing/plot_using_external_data.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        :ref:`Go to the end <sphx_glr_download_auto_examples_training_and_testing_plot_using_external_data.py>`
+        to download the full example code.
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_auto_examples_training_and_testing_plot_using_external_data.py:
+
+
+Using External Test Data
+========================================================================
+
+Let's learn how to use external test data with Fusilli!
+Some guidance can also be found in the :ref:`Data Loading <data-loading>` section of the documentation.
+
+The extra step that we need to take is to provide the paths to the test data files to the functions that create evaluation figures: :class:`~.fusilli.eval.RealsVsPreds.from_new_data`, :class:`~.fusilli.eval.ConfusionMatrix.from_new_data`, :class:`~.fusilli.eval.ModelComparison.from_new_data`.
+
+.. note::
+
+    It is not possible to use external test data with graph-based fusion models.
+
+
+We'll rush through the first few steps of the training and testing process, as they are covered in more detail in the other example notebooks.
+
+.. GENERATED FROM PYTHON SOURCE LINES 18-94
+
+.. code-block:: Python
+
+
+    import matplotlib.pyplot as plt
+    from tqdm.auto import tqdm
+    import os
+
+    from fusilli.data import prepare_fusion_data
+    from fusilli.eval import RealsVsPreds
+    from fusilli.train import train_and_save_models
+    from fusilli.utils.model_chooser import import_chosen_fusion_models
+
+    # sphinx_gallery_thumbnail_number = -1
+
+
+    model_conditions = {
+        "class_name": ["ConcatTabularFeatureMaps"],
+    }
+
+    fusion_models = import_chosen_fusion_models(model_conditions)
+
+    # Regression task
+    prediction_task = "regression"
+
+    # Set the batch size
+    batch_size = 48
+
+    # Setting output directories
+    output_paths = {
+        "losses": "loss_logs/external_data",
+        "checkpoints": "checkpoints/external_data",
+        "figures": "figures/external_data",
+    }
+
+    for dir in output_paths.values():
+        os.makedirs(dir, exist_ok=True)
+
+    # Clearing the loss logs directory (only for the example notebooks)
+    for dir in os.listdir(output_paths["losses"]):
+        # remove files
[…]

[…]
