### 2026-07-30 6a92479
Document new capabilities, add changelog, update contact info
Documents three-tabular-modality support, main_modality/attention_modality
selection, and image transforms - all previously undocumented. Adds
CHANGELOG.md covering everything unreleased since v1.2.3. Updates the
README bio to past tense and points contact queries to GitHub issues
instead of a UCL email address.
