### 2026-07-30 a631284
Fix stale TODOs, wrong license classifier, Python floor, and CI PR trigger
- remove TODO comments for 3-modality/main-modality support that was already
  implemented (activation.py, attention_and_activation.py, crossmodal_att.py,
  concat_feature_maps.py)
- pyproject.toml claimed MIT License in its classifier; the actual LICENSE
  file is AGPLv3, matching the README - classifier was just wrong
- bumped requires-python to >=3.9, matching what numpy~=1.25 and the
  urllib3>=2.5.0 fix actually need (was still >=3.7, which would produce a
  confusing dependency resolution failure rather than a cle
[…]
