+++ fusilli/fusionmodels/tabularfusion/activation.py
@@ -2,11 +2,6 @@
 Activation-function fusion model for tabular data.
 """
 
-# TODO make 3-tabular data work
-
-# TODO add argument to say which tabular modality is the best one
-
-
 import torch.nn as nn
 from fusilli.fusionmodels.base_model import ParentFusionModel
 import torch
