+++ .github/workflows/python-package.yml
@@ -8,7 +8,9 @@ on:
     branches: [ "main" ]
   pull_request:
     types:
-      - closed
+      - opened
+      - synchronize
+      - reopened
 
 jobs:
   build:
