### 2026-07-30 8d6fc18
Swap kfold/regression examples to ConcatTabularData to fix mode collapse
