+++ docs/auto_examples/training_and_testing/plot_one_model_binary_kfold.py
@@ -4,7 +4,7 @@ K-Fold Cross-Validation: Binary Classification
 
 🚀 In this tutorial, we'll explore binary classification using K-fold cross validation. 
 We'll show you how to train a fusion model using K-Fold cross-validation with multimodal tabular data. 
-Specifically, we're using the :class:`~.TabularCrossmodalMultiheadAttention` model.
+Specifically, we're using the :class:`~.ConcatTabularData` model.
 
 Data:
 
@@ -33,12 +33,12 @@ from fusilli.train import train_and_save_models
 # %%
 # 1. Import the fusion model 🔍
 # --------------------------------
-# We're importing only one model for this example, the :class:`~.TabularCrossmodalMultiheadAttention` model.
+# We're importing only one model for this example, the :class:`~.ConcatTabularData` model.
 # Instead of using the :func:`~fusilli.utils.model_chooser.import_chosen_fusion_models` function, we're importing the model directly like with any other library method.
 
 
-from fusilli.fusionmodels.tabularfusion.crossmodal_att import (
-    TabularCrossmodalMultiheadAttention,
+from fusilli.fusionmodels.tabularfusion.concat_data import (
+    ConcatTabularData,
 )
 
 # %%
@@ -118,7 +118,7 @@ data_paths = {
 # We're not using checkpointing for this example, so we set ``enable_checkpointing=False``. We're also setting ``show_loss_plot=True`` to plot the loss curves for each fold.
 
 
-fusion_model = TabularCrossmodalMultiheadAttention
+fusion_model = ConcatTabularData
 
 print("method_name:", fusion_model.method_name)
 print("modality_type:", fusion_model.modality_type)
