### 2026-07-30 e892727
Clarify breaking changes mostly don't affect normal usage
